+++ tests/ethercat/test_pdo.py
@@ -328,6 +328,8 @@ def test_start_stop_pdo(connect_to_all_slave):
     servos, net = connect_to_all_slave
     operation_mode_uid = "DRV_OP_CMD"
     rpdo_registers = [operation_mode_uid]
+    operation_mode_display_uid = "DRV_OP_VALUE"
+    tpdo_registers = [operation_mode_display_uid]
     default_operation_mode = 1
     current_operation_mode = {}
     new_operation_mode = {}
@@ -336,7 +338,7 @@ def test_start_stop_pdo(connect_to_all_slave):
         new_operation_mode[index] = default_operation_mode
         if current_operation_mode[index] == default_operation_mode:
             new_operation_mode[index] += 1
-        rpdo_map, tpdo_map = create_pdo_maps(servo, rpdo_registers, TPDO_REGISTERS)
+        rpdo_map, tpdo_map = create_pdo_maps(servo, rpdo_registers, tpdo_registers)
         for item in rpdo_map.items:
             item.value = new_operation_mode[index]
         servo.set_pdo_map_to_slave([rpdo_map], [tpdo_map])
@@ -344,13 +346,10 @@ def test_start_stop_pdo(connect_to_all_slave):
     for index, servo in enumerate(servos):
         # Check that RPDOs are being received by the slave
         assert servo._rpdo_maps[0].items[0].value == servo.read(operation_mode_uid)
+        # Check that TPDOs are being sent by the slave
+        assert servo._tpdo_maps[0].items[0].value == servo.read(tpdo_registers[0])
         # Restore the previous operation mode
         servo.write(operation_mode_uid, current_operation_mode[index])
-        # Check that TPDOs are being sent by the slave
-        # TODO: Confirm this approx is needed in INGK-839
-        assert pytest.approx(servo._tpdo_maps[0].items[0].value, abs=2) == servo.read(
-            TPDO_REGISTERS[0]
-        )
     # Check that PDOs can be re-started with the same configuration
     start_stop_pdos(net)
     # Re-configure the PDOs and re-start the PDO exchange
@@ -578,3 +577,25 @@ def test_remove_tpdo_map_exceptions(connect_to_slave, create_pdo_map):
         servo.remove_tpdo_map(rpdo_map)
     with pytest.raises(IndexError):
         servo.remove_tpdo_map(tpdo_map_index=1)
+
+
+@pytest.mark.no_connection
+def test_rpdo_map_set_items_bytes(create_pdo_map):
+    _, rpdo_map = create_pdo_map
+    data_bytes = bytes()
+    for idx, item in enumerate(rpdo_map.items):
+        data_bytes += convert_dtype_to_bytes(idx, item.register.dtype)
+    rpdo_map.set_item_bytes(data_bytes)
+    for idx, item in enumerate(rpdo_map.items):
+        assert item.value == idx
+
+
+@pytest.mark.no_connection
+def test_tpdo_map_set_items_bytes(create_pdo_map):
+    tpdo_map, _ = create_pdo_map
+    data_bytes = bytes()
+    for idx, item in enumerate(tpdo_map.items):
+        data_bytes += convert_dtype_to_bytes(idx, item.register.dtype)
+    tpdo_map.set_item_bytes(data_bytes)
+    for idx, item in enumerate(tpdo_map.items):
+        assert item.value == idx